Main module - client details (wip).
git-svn-id: http://svn.apollosoft.net.au:18080/svn/myplanner/trunk@57 65940d0d-0698-491e-b608-6f35d31c8762

+++ core/src/main/java/com/argus/financials/dao/EntityDao.java
@@ -20,6 +20,6 @@ public interface EntityDao extends BaseDAO
      * @param countryId
      * @return
      */
-    List<State> findStates(Long countryId);
+    List<State> findStates(Integer countryId);
 
 }
